requirements.txt
# fastapi
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.7.3

# AI & ML
anthropic==0.18.1
openai>=1.33.0

# Database
qdrant-client>=1.7.3
supabase==2.3.4

# Google Cloud
google-cloud-speech==2.24.0
google-cloud-texttospeech==2.16.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# HTTP clients
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-multipart==0.0.6

# Linear API
gql[all]==3.5.0

# Notion API  
notion-client==2.2.1

# Beeper MCP (if available)
# mcp-beeper (TBD)

# Memory
mem0ai==0.0.8

# Vector operations
numpy==1.26.3

# Websockets (for Aquarium)
websockets==12.0

# Voice - ElevenLabs
elevenlabs>=1.0.0
